refactor: 패키지 정리 및 컨벤션
