+++ nextstep-di/src/main/java/nextstep/di/domain/beandefinition/AnnotationBeanDefinition.java
@@ -1,6 +1,6 @@
-package nextstep.di.factory.domain.beandefinition;
+package nextstep.di.domain.beandefinition;
 
-import nextstep.di.factory.util.ReflectionUtils;
+import nextstep.di.util.ReflectionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
