+++ nextstep-di/src/main/java/nextstep/di/domain/context/ApplicationContext.java
@@ -1,10 +1,12 @@
-package nextstep.di.factory.domain.context;
+package nextstep.di.domain.context;
 
 import java.lang.annotation.Annotation;
 import java.util.Set;
 
 public interface ApplicationContext {
     void initialize();
+
     <T> T getBean(Class<T> requiredType);
+
     Set<Class<?>> getSupportedClassByAnnotation(Class<? extends Annotation> annotation);
 }
