+++ nextstep-di/src/main/java/nextstep/di/domain/context/GenericApplicationContext.java
@@ -1,12 +1,12 @@
-package nextstep.di.factory.domain.context;
+package nextstep.di.domain.context;
 
 import nextstep.annotation.ComponentScan;
 import nextstep.annotation.Configuration;
-import nextstep.di.factory.domain.BeanFactory;
-import nextstep.di.factory.domain.GenericBeanFactory;
-import nextstep.di.factory.domain.scanner.ClassPathScanner;
-import nextstep.di.factory.domain.scanner.ConfigurationScanner;
-import nextstep.di.factory.support.SupportedClass;
+import nextstep.di.domain.factory.BeanFactory;
+import nextstep.di.domain.factory.GenericBeanFactory;
+import nextstep.di.domain.scanner.ClassPathScanner;
+import nextstep.di.domain.scanner.ConfigurationScanner;
+import nextstep.di.support.SupportedClass;
 
 import java.lang.annotation.Annotation;
 import java.util.Set;
