+++ nextstep-di/src/main/java/nextstep/di/domain/factory/BeanFactory.java
@@ -1,11 +1,13 @@
-package nextstep.di.factory.domain;
+package nextstep.di.domain.factory;
 
-import nextstep.di.factory.domain.beandefinition.BeanDefinition;
+import nextstep.di.domain.beandefinition.BeanDefinition;
 
 import java.util.Set;
 
 public interface BeanFactory {
     <T> T getBean(Class<T> requiredType);
+
     void addBeanDefinition(Class<?> clazz, BeanDefinition beanDefinition);
+
     void addInstantiateBeans(Set<Class<?>> preInstantiateBeans);
 }
