+++ nextstep-di/src/main/java/nextstep/di/exception/ReflectionUtilException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.exception;
+package nextstep.di.exception;
 
 public class ReflectionUtilException extends RuntimeException {
     public ReflectionUtilException(String message, Throwable cause) {
