+++ nextstep-di/src/main/java/nextstep/di/util/BeanFactoryUtils.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.util;
+package nextstep.di.util;
 
 import com.google.common.collect.Sets;
 import nextstep.annotation.Inject;
