+++ nextstep-di/src/test/java/nextstep/di/factory/domain/beandefinition/ConstructorBeanTest.java
@@ -1,5 +1,7 @@
 package nextstep.di.factory.domain.beandefinition;
 
+import nextstep.di.domain.beandefinition.AnnotationBeanDefinition;
+import nextstep.di.domain.beandefinition.BeanDefinition;
 import nextstep.di.factory.example.JdbcUserRepository;
 import org.junit.jupiter.api.Test;
