+++ nextstep-di/src/test/java/nextstep/di/factory/support/ReflectionUtilsTest.java
@@ -2,8 +2,8 @@ package nextstep.di.factory.support;
 
 import nextstep.di.factory.example.JdbcUserRepository;
 import nextstep.di.factory.example.MyQnaService;
-import nextstep.di.factory.exception.ReflectionUtilException;
-import nextstep.di.factory.util.ReflectionUtils;
+import nextstep.di.exception.ReflectionUtilException;
+import nextstep.di.util.ReflectionUtils;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
